+++ server/package.json
@@ -17,6 +17,7 @@
     "mongoose": "^7.0.4",
     "node-cron": "^3.0.2",
     "nodemailer": "^6.9.1",
+    "nodemon": "^2.0.22",
     "passport-microsoft": "^1.0.0"
   },
   "description": ""
