+++ server/routes/new.html
@@ -7,6 +7,6 @@
     <title>Document</title>
 </head>
 <body>
-    <h1>Success</h1>
+    <h1>Server is working</h1>
 </body>
 </html>
