+++ server/routes/user_route.js
@@ -1,0 +1,34 @@
+const express = require("express");
+const userModel = require("../models/usermodel");
+
+const app = express();
+
+//*********************user calls***********************
+
+
+//create a user
+
+
+app.post("/add_user", async (request, response) => {
+    const user = new userModel(request.body);
+  
+    try {
+      await user.save();
+      response.send(user);
+    } catch (error) {
+      response.status(500).send(error);
+    }
+});
+
+//retrieve all user
+app.get("/users", async (request, response) => {
+    const users = await userModel.find({});
+  
+    try {
+      response.send(users);
+    } catch (error) {
+      response.status(500).send(error);
+    }
+  });
+
+  module.exports = app;
