+++ server/server.js
@@ -1,13 +1,13 @@
 const express = require('express');
 const mongoose = require('mongoose');
-const Router = require("./routes");
 const cors  = require("cors");
-
-
+const user_route = require("./routes/user_route");
+const dsr_route = require("./routes/dsr_route");
+const draft_route = require("./routes/draft_route");
+const web_route = require("./routes/web_route");
 const cron = require('node-cron');
 const nodemailer = require('nodemailer');
-// const model = require("./model")
-// const jsonData = require('./data.json');
+
 
 // create the Express app
 const app = express();
@@ -87,7 +87,10 @@ cron.schedule('0 22 * * *', () => {
 
 
 // start the server
-app.use(Router);
+app.use(dsr_route);
+app.use(user_route);
+app.use(draft_route);
+app.use(web_route);
 
 
 const PORT = process.env.PORT || 3030;
